// src/app/models/vehicle.ts
export class Vehicle {
  id: number;
  name: string;
  year: string;
  cardesc: string;
  maker: string;
  bodytype: string;
  carmodel: string;
  modeltrim: string;
  avg_rating: number;
  no_of_rating: number;
  msrp: number;
  invoice: number;
  profit: number;
  month: number;
  residual10: number;
  residual: number;
  residual15: number;
  moneyfactor: number;
  leasecash: number;
  bankfee: number;
  downpay: number;
  payments: number;
  apr36: number;
  apr48: number;
  apr: number;
  apr72: number;
  rebate: number;
  loan_payments: number;
}